refactor and restructure

+++ Ledgerly.API/Models/DTOs/Account/RecordAuditLogAccount.cs
@@ -17,7 +17,7 @@ namespace Ledgerly.API.Models.DTOs.Account
         public string? Memo { get; set; }
 
         [Required]
-        public short StatusId { get; set; }
+        public int StatusId { get; set; }
 
         [Required]
         public string StatusName { get; set; } = null!;
